Finish documetation and move fountain module to fountain-core
