+++ .circleci/config.yml
@@ -9,7 +9,7 @@ jobs:
     steps:
       - checkout
       - restore_cache:
-          key: jars-{{ checksum "build.gradle" }}-{{ checksum "versions.gradle" }}-{{ checksum  "fountain/build.gradle" }}-{{ checksum  "app/build.gradle" }}
+          key: jars-{{ checksum "build.gradle" }}-{{ checksum "versions.gradle" }}-{{ checksum  "app/build.gradle" }}-{{ checksum  "fountain-core/build.gradle" }}-{{ checksum  "fountain-rx2/build.gradle" }}
       - run:
           name: Accept licenses
           command: yes | sdkmanager --licenses || true
@@ -19,10 +19,10 @@ jobs:
       - save_cache:
           paths:
             - ~/.gradle
-          key: jars-{{ checksum "build.gradle" }}-{{ checksum "versions.gradle" }}-{{ checksum  "fountain/build.gradle" }}-{{ checksum  "app/build.gradle" }}
+          key: jars-{{ checksum "build.gradle" }}-{{ checksum "versions.gradle" }}-{{ checksum  "app/build.gradle" }}-{{ checksum  "fountain-core/build.gradle" }}-{{ checksum  "fountain-rx2/build.gradle" }}
       - run:
           name: Build Project
-          command: ./gradlew build
+          command: ./gradlew clean build
       - run:
           name: Run detektCheck
           command: ./gradlew detektCheck
