+++ build.gradle
@@ -1,7 +1,6 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    ext.kotlin_version = '1.2.51'
     apply from: 'versions.gradle'
     repositories {
         google()
@@ -17,7 +16,6 @@ buildscript {
         classpath "org.jetbrains.dokka:dokka-android-gradle-plugin:${libraryVersion.dokka}"
         classpath "org.jetbrains.kotlin:kotlin-android-extensions:${libraryVersion.kotlin}"
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:${libraryVersion.kotlin}"
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
     }
 }
