+++ fountain-core/.gitignore
