+++ fountain-core/src/androidTest/java/com/xmartlabs/fountain/ExampleInstrumentedTest.java
