+++ fountain-core/src/main/AndroidManifest.xml
