+++ fountain-core/src/main/java/android/arch/paging/PagingRequestHelper.java
