+++ fountain-core/src/main/java/com/xmartlabs/fountain/NetworkState.kt
