+++ fountain-core/src/main/java/com/xmartlabs/fountain/adapter/CachedDataSourceAdapter.kt
