+++ fountain-core/src/main/java/com/xmartlabs/fountain/feature/cachednetwork/BoundaryCallback.kt
