+++ fountain-core/src/main/res/values/strings.xml
