+++ fountain-core/src/test/java/com/xmartlabs/fountain/common/MockedNetworkDataSourceAdapter.kt
