+++ fountain-core/src/test/java/com/xmartlabs/fountain/networkstate/NetworkStatusUnitTest.kt
