+++ fountain-core/src/test/java/com/xmartlabs/fountain/pagefetcher/totalentities/NetworkDataSourceWithTotalEntityCountAdapterUnitTest.kt
