+++ fountain-core/src/test/java/com/xmartlabs/fountain/pagefetcher/totalentities/RefreshNetworkModeNetworkDataSourceWithTotalEntityCountAdapterUnitTest.kt
