+++ fountain-core/src/test/java/com/xmartlabs/fountain/refresh/RefreshUnitTest.kt
